requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
isort==5.13.2
psycopg2-binary==2.9.9
python-decouple==3.8
ruff==0.6.3
sqlparse==0.5.1
